Moved error combinators into the errors package
